feat: update member activity

+++ src/dto/transaction_dto.rs
@@ -26,9 +26,9 @@ pub struct TransactionDTO{
         deserialize_with = "deserialize_object_id"
     )]
     pub member_id:Option<ObjectId>,
-    pub total_price_before_discount:i64,
+    pub total_price_before_discount:f64,
     pub notes:String,
-    pub total_price:i64,
+    pub total_price:f64,
     pub total_discount:f64,
     pub is_membership:bool,
     #[serde(
